// src/repository/openapi/model/customer-summary-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Trust Score Dashboard API
 * 기업별 신뢰 점수 시계열 데이터 API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { DateRange } from './date-range';

/**
 * 고객 요약 조회 요청
 * @export
 * @interface CustomerSummaryRequest
 */
export interface CustomerSummaryRequest {
    /**
     * 
     * @type {DateRange}
     * @memberof CustomerSummaryRequest
     */
    'dateRange': DateRange;
}