// src/repository/openapi/model/customer-summary-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Trust Score Dashboard API
 * 기업별 신뢰 점수 시계열 데이터 API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { Category } from './category';
// May contain unused imports in some cases
// @ts-ignore
import type { CompanySize } from './company-size';
// May contain unused imports in some cases
// @ts-ignore
import type { CustomerDetailPeriodData } from './customer-detail-period-data';
// May contain unused imports in some cases
// @ts-ignore
import type { ProductType } from './product-type';

/**
 * 고객 요약 응답
 * @export
 * @interface CustomerSummaryResponse
 */
export interface CustomerSummaryResponse {
    /**
     * 
     * @type {number}
     * @memberof CustomerSummaryResponse
     */
    'companyId': number;
    /**
     * 
     * @type {string}
     * @memberof CustomerSummaryResponse
     */
    'companyName': string;
    /**
     * 
     * @type {string}
     * @memberof CustomerSummaryResponse
     */
    'manager'?: string | null;
    /**
     * 
     * @type {Category}
     * @memberof CustomerSummaryResponse
     */
    'category'?: Category | null;
    /**
     * 
     * @type {CompanySize}
     * @memberof CustomerSummaryResponse
     */
    'companySize'?: CompanySize | null;
    /**
     * 
     * @type {Array<ProductType>}
     * @memberof CustomerSummaryResponse
     */
    'productUsage'?: Array<ProductType>;
    /**
     * 
     * @type {number}
     * @memberof CustomerSummaryResponse
     */
    'contractAmount'?: number | null;
    /**
     * 
     * @type {CustomerDetailPeriodData}
     * @memberof CustomerSummaryResponse
     */
    'current': CustomerDetailPeriodData;
    /**
     * 
     * @type {CustomerDetailPeriodData}
     * @memberof CustomerSummaryResponse
     */
    'previous': CustomerDetailPeriodData;
    /**
     * 
     * @type {string}
     * @memberof CustomerSummaryResponse
     */
    'hubspotUrl': string;
}