// src/repository/openapi/model/possibility.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Trust Score Dashboard API
 * 기업별 신뢰 점수 시계열 데이터 API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 계약 가능성 (confirmation_of_sales_secured 매핑)
 * @export
 * @enum {string}
 */

export const Possibility = {
    _100: '100%',
    _90: '90%',
    _40: '40%',
    _10: '10%',
    _0: '0%'
} as const;

export type Possibility = typeof Possibility[keyof typeof Possibility];