// src/repository/openapi/model/sales-history-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Trust Score Dashboard API
 * 기업별 신뢰 점수 시계열 데이터 API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { DateRange } from './date-range';

/**
 * 영업 히스토리 조회 요청
 * @export
 * @interface SalesHistoryRequest
 */
export interface SalesHistoryRequest {
    /**
     * 
     * @type {DateRange}
     * @memberof SalesHistoryRequest
     */
    'dateRange': DateRange;
}