// src/repository/openapi/model/sales-history-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Trust Score Dashboard API
 * 기업별 신뢰 점수 시계열 데이터 API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { SalesAction } from './sales-action';

/**
 * 영업 히스토리 응답
 * @export
 * @interface SalesHistoryResponse
 */
export interface SalesHistoryResponse {
    /**
     * 
     * @type {number}
     * @memberof SalesHistoryResponse
     */
    'companyId': number;
    /**
     * 
     * @type {string}
     * @memberof SalesHistoryResponse
     */
    'companyName': string;
    /**
     * 
     * @type {Array<SalesAction>}
     * @memberof SalesHistoryResponse
     */
    'salesActions': Array<SalesAction>;
}